decimalfield to floatfield

+++ number-plate-app/sweets/models.py
@@ -4,26 +4,11 @@ from rest_framework.reverse import reverse as api_reverse
 
 class SweetsData(models.Model):
     name = models.CharField(max_length=30)
-    serving_grams = models.DecimalField(
-        max_digits=6,
-        decimal_places=2
-    )
-    kcal_100 = models.DecimalField(
-        max_digits=6,
-        decimal_places=2
-    )
-    fat_100 = models.DecimalField(
-        max_digits=6,
-        decimal_places=2
-    )
-    carbohydrate_100 = models.DecimalField(
-        max_digits=6,
-        decimal_places=2
-    )
-    protein_100 = models.DecimalField(
-        max_digits=6,
-        decimal_places=2
-    )
+    serving_grams = models.FloatField()
+    kcal_100 = models.FloatField()
+    fat_100 = models.FloatField()
+    carbohydrate_100 = models.FloatField()
+    protein_100 = models.FloatField()
     timestamp = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
